Task5 simulates user interaction appium/appium_script.py
# appium_script.py

from appium import webdriver
from config import CAPABILITIES
from appium.webdriver.common.appiumby import AppiumBy

def main():
    # Initialize the Appium driver
    driver = webdriver.Remote('http://localhost:4723/wd/hub', CAPABILITIES)

    try:
        # Simulate user login
        driver.find_element(AppiumBy.ID, 'com.example.app:id/username').send_keys('testuser')
        driver.find_element(AppiumBy.ID, 'com.example.app:id/password').send_keys('password')
        driver.find_element(AppiumBy.ID, 'com.example.app:id/login_button').click()

        # Navigate through the app
        driver.find_element(AppiumBy.ID, 'com.example.app:id/menu_button').click()
        driver.find_element(AppiumBy.ID, 'com.example.app:id/settings').click()

        # Data submission
        driver.find_element(AppiumBy.ID, 'com.example.app:id/data_field').send_keys('Some data')
        driver.find_element(AppiumBy.ID, 'com.example.app:id/submit_button').click()

        # Handle screen orientation
        driver.orientation = 'LANDSCAPE'
        driver.orientation = 'PORTRAIT'

    finally:
        # Clean up and close the driver
        driver.quit()

if __name__ == '__main__':
    main()
