Task5 simulates user interaction appium/requirements.txt
Appium-Python-Client
selenium
